requirements.txt
# requirements.txt for comprehensive Spanish generator
openai>=1.0.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.5.0
numpy>=1.24.0
tqdm>=4.64.0

# Optional: for advanced formatting
python-docx>=0.8.11
markdown>=3.4.0
